working on multiple lines

+++ main.c
@@ -184,7 +184,8 @@ void buffer_delete(void) {
 
 void render_cursor(SDL_Renderer *renderer, const Font *font) {
 
-  const Vec2f pos = vec2f(buffer_cursor * FONT_CHAR_WIDTH * FONT_SCALE, 0.0f);
+  const Vec2f pos =
+      vec2f((float)buffer_cursor * FONT_CHAR_WIDTH * FONT_SCALE, 0.0f);
 
   const SDL_Rect rect = {
       .x = (int)floorf(buffer_cursor * FONT_CHAR_WIDTH * FONT_SCALE),
@@ -203,11 +204,11 @@ void render_cursor(SDL_Renderer *renderer, const Font *font) {
   }
 }
 
-// TODO: jump forward/backward by word
+// TODO: Multiple lines
+// TODO: Save/Load file
+// TODO: Jump forward/backward by word
 // TODO: Delete a word
-// TODO: Blinking cursor,
-// TODO: Multiple lines,
-// TODO: Save/Load file.
+// TODO: Blinking cursor
 
 int main(int argc, char **argv) {
